Fix type of size parameter

+++ lib/widget/paragraph.h
@@ -66,7 +66,7 @@ private:
 	ParagraphState renderState;
 
 	void updateLayout();
-	void resizeLines(uint32_t size);
+	void resizeLines(size_t size);
 
 	friend class ParagraphLine;
 };
